rename opening book data file for case-insensitive filesystems
